fix：修改删除司机时，出现账户密码误删的情况。

+++ ses-service/ses-service-base/ses-service-foundation/src/main/java/com/redescooter/ses/service/foundation/service/impl/AccountBaseServiceImpl.java
@@ -13,15 +13,7 @@ import com.redescooter.ses.api.common.enums.tenant.TenantStatusEnum;
 import com.redescooter.ses.api.common.enums.user.UserEventEnum;
 import com.redescooter.ses.api.common.enums.user.UserStatusEnum;
 import com.redescooter.ses.api.common.vo.CountByStatusResult;
-import com.redescooter.ses.api.common.vo.base.BaseCustomerResult;
-import com.redescooter.ses.api.common.vo.base.BaseMailTaskEnter;
-import com.redescooter.ses.api.common.vo.base.BaseUserResult;
-import com.redescooter.ses.api.common.vo.base.BooleanResult;
-import com.redescooter.ses.api.common.vo.base.DateTimeParmEnter;
-import com.redescooter.ses.api.common.vo.base.GeneralEnter;
-import com.redescooter.ses.api.common.vo.base.GeneralResult;
-import com.redescooter.ses.api.common.vo.base.IdEnter;
-import com.redescooter.ses.api.common.vo.base.SetPasswordEnter;
+import com.redescooter.ses.api.common.vo.base.*;
 import com.redescooter.ses.api.foundation.exception.FoundationException;
 import com.redescooter.ses.api.foundation.service.MailMultiTaskService;
 import com.redescooter.ses.api.foundation.service.base.AccountBaseService;
@@ -44,13 +36,7 @@ import com.redescooter.ses.service.foundation.dao.base.PlaTenantMapper;
 import com.redescooter.ses.service.foundation.dao.base.PlaUserMapper;
 import com.redescooter.ses.service.foundation.dao.base.PlaUserPasswordMapper;
 import com.redescooter.ses.service.foundation.dao.base.PlaUserPermissionMapper;
-import com.redescooter.ses.service.foundation.dm.base.PlaTenant;
-import com.redescooter.ses.service.foundation.dm.base.PlaTenantConfig;
-import com.redescooter.ses.service.foundation.dm.base.PlaTenantNode;
-import com.redescooter.ses.service.foundation.dm.base.PlaUser;
-import com.redescooter.ses.service.foundation.dm.base.PlaUserNode;
-import com.redescooter.ses.service.foundation.dm.base.PlaUserPassword;
-import com.redescooter.ses.service.foundation.dm.base.PlaUserPermission;
+import com.redescooter.ses.service.foundation.dm.base.*;
 import com.redescooter.ses.service.foundation.exception.ExceptionCodeEnums;
 import com.redescooter.ses.service.foundation.service.base.PlaTenantConfigService;
 import com.redescooter.ses.service.foundation.service.base.PlaTenantNodeService;
@@ -72,13 +58,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.transaction.annotation.Transactional;
 import redis.clients.jedis.JedisCluster;
 
-import java.util.ArrayList;
-import java.util.Date;
-import java.util.HashMap;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Map;
-import java.util.Set;
+import java.util.*;
 import java.util.stream.Collectors;
 
 /**
@@ -810,14 +790,18 @@ public class AccountBaseServiceImpl implements AccountBaseService {
         }
         userMapper.deleteById(enter.getId());
 
-        QueryWrapper<PlaUserPassword> query = new QueryWrapper<>();
-        query.eq(PlaUserPassword.COL_LOGIN_NAME, plaUser.getLoginName());
-        query.eq(PlaUserPassword.COL_DR, 0);
-        PlaUserPassword userPassword = userPasswordMapper.selectOne(query);
-        if (userPassword == null) {
-            return new GeneralResult(enter.getRequestId());
+        //司机离职  一个邮箱绑定一个账户的时候才能删除密码
+        if (userMapper.selectCount(new LambdaQueryWrapper<PlaUser>().eq(PlaUser::getLoginName, plaUser.getLoginName()).eq(PlaUser::getDr, 0)) == 0) {
+            QueryWrapper<PlaUserPassword> query = new QueryWrapper<>();
+            query.eq(PlaUserPassword.COL_LOGIN_NAME, plaUser.getLoginName());
+            query.eq(PlaUserPassword.COL_DR, 0);
+            PlaUserPassword userPassword = userPasswordMapper.selectOne(query);
+            if (userPassword == null) {
+                return new GeneralResult(enter.getRequestId());
+            }
+            userPasswordMapper.deleteById(userPassword.getId());
         }
-        userPasswordMapper.deleteById(userPassword.getId());
+
 
         QueryWrapper<PlaUserPermission> wrapper = new QueryWrapper<>();
         wrapper.eq(PlaUserPermission.COL_USER_ID, plaUser.getId());
@@ -838,7 +822,7 @@ public class AccountBaseServiceImpl implements AccountBaseService {
      * @param enter
      * @return
      */
-    @Transactional
+    @Transational
     @Override
     public GeneralResult sendEmailActiv(IdEnter enter) {
